Add task to regather facts once the role completes
In order to ensure the deployment host has the most up-to-date facts we
must regather them upon the completion of the role. This is done because
the role will run several tasks that can perform a state change on
various resources and we need to capture the most current state. This is
especially important when running in an environment that makes use of
cached facts.

Change-Id: I429485eba7e4e6643d10d0f3826471f47aab87fc
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>