Tuneup prep and local routing when running w/out bridges
When running without bridges the local routing needs to be scoped to the
host. This change ensures that is the care.

When re-preping an environment the GPG setup will require aceess to fd's
which may not exist at the time the base cache is re-prep'd.

nspawn mtu settings will now follow the primary interface.

Change-Id: I74e9301a98cf92161feb31e6808e9e02a07f662c
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>