Change ssh variable in Ubuntu and remove masks
The ssh service on ubuntu based systems is "ssh" which is established by
the service unit path `/lib/systemd/system/ssh.service`. When running
the service will respond to the name "sshd" however this is just an
alias.

The image prep task used to mask ssh after it was disabled. While
in practice thats OK, its really a little overkill and provides little
benifit. In prep the containers will have ssh disabled, but in the event
its needed it can be easily enabled wihout having to remember to unmask. 

Change-Id: I2edb3ca7cae6e05f15ab29b1d50ff1281762905e
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>