Ensure /tmp is writeable during cache preparation
This is necessary to ensure it is possible for apt to write to /tmp
when accessing the container image as a chroot during cache prep.

In addition fix up cross-repo zuul jobs to be in line with the rest of OSA
by dropping the Suse job and switching debian stable(stretch) for buster.

Temporarily make the cross-repo jobs non-voting to break a circular
dependancy between fixes in nspawn_hosts and nspawn_container_create.

Depends-On: https://review.opendev.org/728995
Change-Id: I8c6034f49600b9f1f431eb2a7c02ae567ea6d32c